Updated post table
Displays up to 10 posts from the database onto the main page.
