+++ index.php
@@ -1,6 +1,5 @@
 <?php 
 require_once("action/indexAction.php");
-
 	$action = new indexAction();
 	$action->execute();
 require_once("partial/header.php") 
@@ -8,6 +7,9 @@ require_once("partial/header.php")
 
 
 <?php
+require_once("action/postTableAction.php");
+    $action = new postTableAction();
+    $action->execute();
 require_once("Post Table.php");
 ?>
